feat: 타이머 애니메이션 로직을 별도 훅으로 분리
